fix: cleanup inpage scripts
