+++ src/extension/inpage-script/alby.js
@@ -1,4 +1,4 @@
-import AlbyProvider from "../ln/alby";
+import AlbyProvider from "../providers/alby";
 
 if (document) {
   window.alby = new AlbyProvider();
