--- src/extension/ln/alby/index.ts
@@ -1,79 +1,0 @@
-export default class AlbyProvider {
-  enabled: boolean;
-
-  constructor() {
-    this.enabled = false;
-  }
-
-  enable() {
-    if (this.enabled) {
-      return Promise.resolve({ enabled: true });
-    }
-    return this.execute("enable").then((result) => {
-      if (typeof result.enabled === "boolean") {
-        this.enabled = result.enabled;
-      }
-      return result;
-    });
-  }
-
-  /**
-   * Adds a wallet to the user's Alby extension
-   *
-   * @param name The name of the account
-   * @param connector The connector to use
-   * @param config The config to pass to the connector
-   * @returns Nothing, throws if user rejects
-   */
-  addAccount(name: string, connector: string, config: Record<string, unknown>) {
-    if (!this.enabled) {
-      throw new Error("Provider must be enabled before calling addAccount");
-    }
-    return this.execute("addAccount", { connector, name, config });
-  }
-
-  execute(
-    action: string,
-    args?: Record<string, unknown>
-  ): Promise<Record<string, unknown>> {
-    return new Promise((resolve, reject) => {
-      // post the request to the content script. from there it gets passed to the background script and back
-      // in page script can not directly connect to the background script
-      window.postMessage(
-        {
-          application: "LBE",
-          prompt: true,
-          action: `alby/${action}`,
-          scope: "alby",
-          args,
-        },
-        "*" // TODO use origin
-      );
-
-      function handleWindowMessage(messageEvent: MessageEvent) {
-        // check if it is a relevant message
-        // there are some other events happening
-        if (
-          !messageEvent.data ||
-          !messageEvent.data.response ||
-          messageEvent.data.application !== "LBE" ||
-          messageEvent.data.scope !== "alby"
-        ) {
-          return;
-        }
-        if (messageEvent.data.data.error) {
-          reject(new Error(messageEvent.data.data.error));
-        } else {
-          // 1. data: the message data
-          // 2. data: the data passed as data to the message
-          // 3. data: the actual response data
-          resolve(messageEvent.data.data.data);
-        }
-        // For some reason must happen only at the end of this function
-        window.removeEventListener("message", handleWindowMessage);
-      }
-
-      window.addEventListener("message", handleWindowMessage);
-    });
-  }
-}
